Activity Overrides
This is a massive commit that touched a lot of files.  The goal of
this commit is to allow the user to disable an action and all of it's
components.

These changes required that every action have it's own class, thus
the addition of the Illumination file.  Each action also has it's
own configuration which must exist on the context.

Also is the ability to "touch" a component.  The usage is up to the
implementer, but the idea is to perform some custom event (e.g. starting
an irrigation).

Other changes include minor functional reorganizations. Notably, an
override method has been given to components and the zone's main
execution method has been reorganized to apply activity overrides.